Move merge_layers to compose
