+++ src/psd_tools/__init__.py
@@ -1,3 +1,4 @@
 from __future__ import absolute_import
 from psd_tools.user_api.psd_image import PSDImage
+from psd_tools.user_api.composer import compose
 from psd_tools.version import __version__
